+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [Unreleased]
 
+## [0.6.9] - 2026-03-23
+
 ### Added
 
 - Add `FinderColumn.finderColumn(String identifier, String header)`
@@ -811,7 +813,8 @@ Stay tuned for more to come...
 - for dependency upgrades
 -->
 
-[Unreleased]: https://github.com/patternfly-java/patternfly-java/compare/v0.6.8...HEAD
+[Unreleased]: https://github.com/patternfly-java/patternfly-java/compare/v0.6.9...HEAD
+[0.6.9]: https://github.com/patternfly-java/patternfly-java/compare/v0.6.8...v0.6.9
 [0.6.8]: https://github.com/patternfly-java/patternfly-java/compare/v0.6.7...v0.6.8
 [0.6.7]: https://github.com/patternfly-java/patternfly-java/compare/v0.6.6...v0.6.7
 [0.6.6]: https://github.com/patternfly-java/patternfly-java/compare/v0.6.5...v0.6.6